+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 *   Added histogram-based plugins (author Axel Jacobsen).
 *   Updated pyqtgraph
 *   Added RangeToolInvocation methods marker_single_add and marker_dual_add.
+*   Added check and UI status warning when device does not support dual markers.
 
 
 ## 0.5.1